interscript js version tests complete
