+++ .gitignore
@@ -4,3 +4,6 @@
 pkg/
 .idea
 .idea/*
+
+# Dependency directories
+node_modules/
