--- node_test/.gitignore
@@ -1,5 +1,0 @@
-node_modules
-build
-npm-debug.log
-.env
-.DS_Store
