--- node_test/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "node_test",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "mocha": "^8.1.3"
-  }
-}
