+++ package.json
@@ -1,0 +1,29 @@
+{
+  "name": "interscript",
+  "version": "1.0.0",
+  "description": "Javascript version of Interscript gem",
+  "main": "index.js",
+  "directories": {
+    "test": "test"
+  },
+  "dependencies": {},
+  "devDependencies": {
+    "mocha": "^8.1.3"
+  },
+  "scripts": {
+    "test": "mocha"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/webdev778/interscript.git"
+  },
+  "keywords": [
+    "interscript"
+  ],
+  "author": "Jaycee Yumol",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/webdev778/interscript/issues"
+  },
+  "homepage": "https://github.com/webdev778/interscript#readme"
+}
